Restructured the configuration for all plugins
