+++ nvim/.config/nvim/init.lua
@@ -1,4 +1,4 @@
 require 'core'
-require 'plugins'
+require 'plugin_manager'
 -- The line beneath this is called `modeline`. See `:help modeline`
 -- vim: ts=2 sts=2 sw=2 et
