--- nvim/.config/nvim/lua/configs/catppuccin.lua
@@ -1,6 +1,0 @@
-return function()
-  require('catppuccin').setup {
-    flavour = 'mocha',
-  }
-  vim.cmd.colorscheme 'catppuccin'
-end
