--- nvim/.config/nvim/lua/configs/oil.lua
@@ -1,4 +1,0 @@
-return function()
-  require('oil').setup {}
-  vim.keymap.set('n', '-', '<CMD>Oil<CR>', { desc = 'Open parent directory' })
-end
