--- nvim/.config/nvim/lua/configs/treesitter.lua
@@ -1,26 +1,0 @@
-return function()
-  require('nvim-treesitter.configs').setup {
-    ensure_installed = { 'bash', 'c', 'diff', 'html', 'lua', 'luadoc', 'markdown', 'markdown_inline', 'query', 'vim', 'vimdoc' },
-    sync_install = false,
-    -- Automatically install missing parsers
-    auto_install = true,
-    ignore_install = {},
-    highlight = {
-      enable = true,
-      -- Some languages depend on vim's regex highlighting system for additional functionality
-      additional_vim_regex_highlighting = { 'ruby' },
-    },
-    indent = {
-      enable = true,
-      disable = { 'ruby' },
-    },
-
-    -- There are additional nvim-treesitter modules that you can use to interact
-    -- with nvim-treesitter. You should go explore a few and see what interests you:
-    --
-    --    - Incremental selection: Included, see `:help nvim-treesitter-incremental-selection-mod`
-    --    - Show your current context: https://github.com/nvim-treesitter/nvim-treesitter-context
-    --    - Treesitter + textobjects: https://github.com/nvim-treesitter/nvim-treesitter-textobjects
-    modules = {},
-  }
-end
