+++ nvim/.config/nvim/lua/plugin_manager/init.lua
@@ -21,7 +21,7 @@ vim.opt.rtp:prepend(lazypath)
 --    :Lazy update
 --
 -- NOTE: Here is where you install your plugins.
-require('lazy').setup(require 'plugins.plugins', {
+require('lazy').setup(require 'plugin_manager.plugins', {
   ui = {
     -- If you are using a Nerd Font: set icons to an empty table which will use the
     -- default lazy.nvim defined Nerd Font icons, otherwise define a unicode icons table
