+++ nvim/.config/nvim/lua/plugins/catppuccin.lua
@@ -1,0 +1,11 @@
+return {
+  'catppuccin/nvim',
+  name = 'catppuccin',
+  priority = 1000,
+  config = function()
+    require('catppuccin').setup {
+      flavour = 'mocha',
+    }
+    vim.cmd.colorscheme 'catppuccin'
+  end,
+}
