+++ nvim/.config/nvim/lua/plugins/luvit-meta.lua
@@ -1,0 +1,1 @@
+return { 'Bilal2453/luvit-meta', lazy = true }
