+++ nvim/.config/nvim/lua/plugins/oil.lua
@@ -1,0 +1,9 @@
+return {
+  'stevearc/oil.nvim',
+  dependencies = { 'nvim-tree/nvim-web-devicons' },
+
+  config = function()
+    require('oil').setup {}
+    vim.keymap.set('n', '-', '<CMD>Oil<CR>', { desc = 'Open parent directory' })
+  end,
+}
