+++ nvim/.config/nvim/lua/plugins/todo-comments.lua
@@ -1,0 +1,13 @@
+return {
+  'folke/todo-comments.nvim',
+  event = 'VimEnter',
+  dependencies = { 'nvim-lua/plenary.nvim' },
+  opts = { signs = false },
+  keys = {
+    {
+      '<leader>st',
+      ':TodoTelescope<CR>',
+      desc = '[S]earch [T]odos',
+    },
+  },
+}
